// src/features/todo/p0019/types.ts
export interface Profile {
  id: string; // should match user.id
  created_at?: string;
  updated_at: string | null;
  last_sign_in_at?: string | null;
  full_name: string | null;
  avatar_url: string | null;
  email?: string | null; // Added email field
  role: 'admin' | 'manager' | 'employee';
  default_project_id?: number | null;
}

export interface TaskAttachment {
  id: number;
  file_name: string;
  file_path: string;
  file_type: string;
  file_size: number;
  created_at: string;
}

export interface TimeLog {
  id: number;
  task_id: number;
  user_id: string;
  start_time: string;
  end_time: string | null;
}

export interface TaskComment {
  id: number;
  created_at: string;
  task_id: number;
  user_id: string;
  content: string;
  profiles: Profile;
}

export interface Project {
  id: number;
  created_at: string;
  name: string;
  created_by: string;
  project_members?: { count: number }[];
  color?: string | null;
}

export interface ProjectMember {
    project_id: number;
    user_id: string;
    created_at: string;
    projects: Project | null;
    profiles?: Profile;
}

export interface MemberDetails extends ProjectMember {
    profiles: Profile;
    task_count: number;
}

export interface TaskAssignee {
    user_id: string;
    completed_at: string | null;
}

export interface Task {
  id: number;
  user_id: string;
  created_by: string;
  created_at: string;
  updated_at: string;
  title: string;
  description: string | null;
  status: 'todo' | 'inprogress' | 'done' | 'cancelled';
  priority: 'low' | 'medium' | 'high';
  due_date: string | null;
  project_id?: number | null;
  project_ids?: number[]; // Support multiple projects
  assignees?: TaskAssignee[]; // Support multiple assignees with status
  projects?: Project; // For showing project info
  task_attachments?: TaskAttachment[];
  assignee?: Profile; // For showing assignee info
  creator?: Profile; // For showing creator info
  task_time_logs?: TimeLog[];
  task_comments?: TaskComment[];
}